One simulation comprises of: 60 trials of learning the 2 familiar cues; 20 trials of testing if systems reaction to the 2 familiar cues when GPi is activated and there is no more learning; 20 trials of testing of the 2 unfamiliar cues with GPi; 20 trials of testing the 2 familiar ones without GPi; 20 trials of testing the 2 unfa- miliar ones without GPi.
There were some simulations that during the learning part, the system learned the wrong
cue as preferable. Some others that there was more cortical learning than striatal, so
the response time was faster without GPi. Because we have Hebbian learning for the cortex
if the system learns something can't forget it, so we have to make sure that BG are
faster than Cortex.